--- app/src/main/java/com/belatrixsf/allstars/ui/home/EmployeePresenter.java
@@ -1,69 +1,0 @@
-/* The MIT License (MIT)
-* Copyright (c) 2016 BELATRIX
-* Permission is hereby granted, free of charge, to any person obtaining a copy
-* of this software and associated documentation files (the "Software"), to deal
-* in the Software without restriction, including without limitation the rights
-* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-* copies of the Software, and to permit persons to whom the Software is
-* furnished to do so, subject to the following conditions:
-
-* The above copyright notice and this permission notice shall be included in all
-* copies or substantial portions of the Software.
-
-* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-* SOFTWARE.
-*/
-package com.belatrixsf.allstars.ui.home;
-
-import com.belatrixsf.allstars.entities.Employee;
-import com.belatrixsf.allstars.managers.EmployeeManager;
-import com.belatrixsf.allstars.ui.common.AllStarsPresenter;
-import com.belatrixsf.allstars.utils.AllStarsCallback;
-import com.belatrixsf.allstars.utils.ServiceError;
-
-import javax.inject.Inject;
-
-/**
- * Created by PedroCarrillo on 4/15/16.
- */
-public class EmployeePresenter extends AllStarsPresenter<EmployeeView> {
-
-    protected EmployeeManager employeeManager;
-    protected Employee employee;
-
-    @Inject
-    public EmployeePresenter(EmployeeView view, EmployeeManager employeeManager) {
-        super(view);
-        this.employeeManager = employeeManager;
-    }
-
-    public void loadEmployeeAccount() {
-        employeeManager.getLoggedInEmployee(new AllStarsCallback<Employee>() {
-            @Override
-            public void onSuccess(Employee employee) {
-                EmployeePresenter.this.employee = employee;
-                showEmployeeData();
-            }
-
-            @Override
-            public void onFailure(ServiceError serviceError) {
-                showError(serviceError.getErrorMessage());
-            }
-        });
-    }
-
-    private void showEmployeeData() {
-        if (employee.getFirstName() != null || employee.getLastName() != null) {
-            view.showEmployeeName(employee.getFullName());
-        }
-        if (employee.getRole() != null) {
-            view.showRole(employee.getRole().getName());
-        }
-    }
-
-}
