--- app/src/main/java/com/belatrixsf/allstars/utils/di/components/EmployeeComponent.java
@@ -1,42 +1,0 @@
-/* The MIT License (MIT)
-* Copyright (c) 2016 BELATRIX
-* Permission is hereby granted, free of charge, to any person obtaining a copy
-* of this software and associated documentation files (the "Software"), to deal
-* in the Software without restriction, including without limitation the rights
-* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-* copies of the Software, and to permit persons to whom the Software is
-* furnished to do so, subject to the following conditions:
-
-* The above copyright notice and this permission notice shall be included in all
-* copies or substantial portions of the Software.
-
-* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-* SOFTWARE.
-*/
-package com.belatrixsf.allstars.utils.di.components;
-
-import com.belatrixsf.allstars.ui.home.EmployeePresenter;
-import com.belatrixsf.allstars.ui.home.EmployeeView;
-import com.belatrixsf.allstars.utils.di.modules.presenters.EmployeePresenterModule;
-import com.belatrixsf.allstars.utils.di.scopes.UIScope;
-
-import dagger.Component;
-
-/**
- * Created by PedroCarrillo on 4/15/16.
- */
-@UIScope
-@Component(
-        dependencies = ApplicationComponent.class,
-        modules = EmployeePresenterModule.class
-)
-public interface EmployeeComponent {
-
-    EmployeePresenter employeePresenter();
-
-}
