+++ app/src/main/res/layout/fragment_account.xml
@@ -6,27 +6,64 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <LinearLayout
-        android:orientation="vertical"
+    <android.support.design.widget.CoordinatorLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+        android:layout_height="match_parent">
 
-        <android.support.v7.widget.CardView
+        <LinearLayout
+            android:orientation="vertical"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="@dimen/activity_horizontal_margin"
-            app:cardBackgroundColor="@android:color/white"
-            app:cardCornerRadius="2dp"
-            app:cardPreventCornerOverlap="false"
-            app:cardUseCompatPadding="true"
-            android:layout_marginLeft="@dimen/activity_horizontal_margin"
-            android:layout_marginRight="@dimen/activity_horizontal_margin">
+            android:layout_height="wrap_content">
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="200dp"
+                android:background="@android:color/white"
+                android:orientation="vertical">
+
+                <View
+                    android:id="@+id/belatrix_placeholder"
+                    android:layout_width="match_parent"
+                    android:background="@color/colorPrimaryLight"
+                    android:layout_height="0dp"
+                    android:layout_weight=".4"/>
+
+                <LinearLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="0dp"
+                    android:layout_weight=".6"
+                    android:gravity="center"
+                    android:orientation="vertical">
+
+                    <TextView
+                        android:id="@+id/profile_name"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginTop="@dimen/dimen_1_5"
+                        style="@android:style/TextAppearance.Large"
+                        tools:text="Pedro Carrillo" />
+
+                    <TextView
+                        android:id="@+id/profile_role"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginTop="@dimen/dimen_1_5"
+                        tools:text="Senior Android Developer" />
+
+                </LinearLayout>
+
+            </LinearLayout>
+
+            <View
+                android:layout_width="match_parent"
+                android:layout_height="1dp"
+                android:background="@color/divider"/>
 
             <LinearLayout
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:orientation="horizontal"
                 android:background="@android:color/white"
+                android:orientation="horizontal"
                 android:weightSum="12">
 
                 <LinearLayout
@@ -40,7 +77,7 @@
                     <TextView
                         android:layout_height="match_parent"
                         android:layout_width="wrap_content"
-                        android:textColor="@color/divider"
+                        android:textColor="@color/subTitle"
                         android:text="@string/title_skype_id"/>
 
                     <TextView
@@ -53,8 +90,6 @@
 
                 </LinearLayout>
                 <View
-                    android:layout_marginTop="10dp"
-                    android:layout_marginBottom="10dp"
                     android:layout_width="1dp"
                     android:layout_height="match_parent"
                     android:background="@color/divider"/>
@@ -69,7 +104,7 @@
                     <TextView
                         android:layout_height="match_parent"
                         android:layout_width="wrap_content"
-                        android:textColor="@color/divider"
+                        android:textColor="@color/subTitle"
                         android:text="@string/title_score"/>
 
                     <TextView
@@ -81,9 +116,8 @@
                         tools:text="1233"/>
 
                 </LinearLayout>
+
                 <View
-                    android:layout_marginTop="10dp"
-                    android:layout_marginBottom="10dp"
                     android:layout_width="1dp"
                     android:layout_height="match_parent"
                     android:background="@color/divider"/>
@@ -99,7 +133,7 @@
                     <TextView
                         android:layout_height="match_parent"
                         android:layout_width="wrap_content"
-                        android:textColor="@color/divider"
+                        android:textColor="@color/subTitle"
                         android:text="@string/title_level"/>
 
                     <TextView
@@ -113,25 +147,27 @@
                 </LinearLayout>
 
             </LinearLayout>
-        </android.support.v7.widget.CardView>
 
-        <android.support.v7.widget.CardView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:cardBackgroundColor="@android:color/white"
-            app:cardCornerRadius="2dp"
-            app:cardPreventCornerOverlap="false"
-            app:cardUseCompatPadding="true"
-            android:layout_marginTop="@dimen/activity_horizontal_margin"
-            android:layout_marginLeft="@dimen/activity_horizontal_margin"
-            android:layout_marginRight="@dimen/activity_horizontal_margin">
+            <View
+                android:layout_width="match_parent"
+                android:layout_height="1dp"
+                android:background="@color/divider"/>
+
             <android.support.v7.widget.RecyclerView
                 android:id="@+id/rv_account_recommendations"
                 android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:background="#F5F5F5"/>
-        </android.support.v7.widget.CardView>
-    </LinearLayout>
+                android:layout_height="match_parent"/>
 
+        </LinearLayout>
+
+        <ImageView
+            android:id="@+id/profile_picture"
+            android:layout_width="70dp"
+            android:layout_height="70dp"
+            app:layout_anchor="@+id/belatrix_placeholder"
+            app:layout_anchorGravity="bottom|center"
+            android:background="@mipmap/ic_launcher"/>
+
+    </android.support.design.widget.CoordinatorLayout>
 
 </android.support.v4.widget.NestedScrollView>
